fix/separation: split main into API classes and implemented generate right-click
